Rename the example projects
Rename TestLib to Library_Generate and StaticLibrary_Example to
Library_Use.

 * Set Library_Generate default target as lib
 * Eliminate relative include paths
 * Add missing newlines
 * Simplify example application
 * Update documentation
