+++ Examples/MAX32650/Library_Generate/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>StaticLibrary_Example</name>
+	<name>Library_Generate</name>
 	<comment></comment>
 	<projects>
 	</projects>
