+++ Examples/MAX32650/Library_Generate/.settings/org.eclipse.cdt.codan.core.prefs
