+++ Examples/MAX32650/Library_Generate/.vscode/flash.gdb
