+++ Examples/MAX32650/Library_Generate/.vscode/tasks.json
@@ -23,38 +23,6 @@
             "problemMatcher": []
         },
         {
-            "label": "flash",
-            "type": "shell",
-            "command": "arm-none-eabi-gdb",
-            "args": [
-                "--cd=\"${workspaceFolder}\"",
-                "--se=\"build/${config:program_file}\"",
-                "--symbols=build/${config:symbol_file}",
-                "-x=\"${workspaceFolder}/.vscode/flash.gdb\"",
-                "--ex=\"flash_m4 ${config:OCD_path} ${config:M4_OCD_interface_file} ${config:M4_OCD_target_file}\"",
-                "--batch"
-            ],
-            "group": "build",
-            "problemMatcher": [],
-            "dependsOn":["build"]
-        },
-        {
-            "label": "flash & run",
-            "type": "shell",
-            "command": "arm-none-eabi-gdb",
-            "args": [                
-                "--cd=\"${workspaceFolder}\"",
-                "--se=\"build/${config:program_file}\"",
-                "--symbols=build/${config:symbol_file}",
-                "-x=\"${workspaceFolder}/.vscode/flash.gdb\"",                
-                "--ex=\"flash_m4_run ${config:OCD_path} ${config:M4_OCD_interface_file} ${config:M4_OCD_target_file}\"",
-                "--batch"
-            ],
-            "group": "build",
-            "problemMatcher": [],
-            "dependsOn":["build"]
-        },
-        {
             "label": "erase flash",
             "type": "shell",
             "command": "openocd",
