+++ Examples/MAX32650/Library_Generate/README.md
@@ -1,0 +1,6 @@
+## Description
+
+Example project that demonstrates how to generate a static library using MSDK.
+
+The output of this project is a ".a" static library file that can be linked by external applications. 
+Refer to Library_Use project for an example implementation that links to this project.
