+++ Examples/MAX32650/Library_Use/.cproject
@@ -61,12 +61,12 @@
 		</cconfiguration>
 	</storageModule>
 	<storageModule moduleId="cdtBuildSystem" version="4.0.0">
-		<project id="StaticLibrary.null.923778415" name="TestLib"/>
+		<project id="Library_Use.null.923778415" name="Library_Use"/>
 	</storageModule>
 	<storageModule moduleId="org.eclipse.cdt.core.LanguageSettingsProviders"/>
 	<storageModule moduleId="refreshScope" versionNumber="2">
 		<configuration configurationName="Default">
-			<resource resourceType="PROJECT" workspacePath="/TestLib"/>
+			<resource resourceType="PROJECT" workspacePath="/Library_Use"/>
 		</configuration>
 	</storageModule>
 	<storageModule moduleId="org.eclipse.cdt.make.core.buildtargets"/>
