+++ Examples/MAX32650/Library_Use/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>TestLib</name>
+	<name>Library_Use</name>
 	<comment></comment>
 	<projects>
 	</projects>
