+++ Examples/MAX32650/Library_Use/.settings/org.eclipse.cdt.codan.core.prefs
