+++ Examples/MAX32650/Library_Use/.vscode/settings.json
