--- Examples/MAX32650/StaticLibrary_Example/README.md
@@ -1,35 +1,0 @@
-## Description
-
-Simple program that demonstrates how to build and link static libraries.
-
-The demo application reads the push button inputs and toggles the LEDs based on the input states. The GPIO
-functions are built into a static library that can be found under TestLib project. Refer to project.mk to
-see how to link your application to a static library.
-
-## Required Connections
-
-### MAX32650 Evaluation Kit
--   Connect a USB cable between the PC and the CN2 (USB TO UART) connector.
--   Connect a debugger to J3 or J4 (JTAG/SWD).
--   Select RX SEL and TX SEL on headers JP12 and JP13.
--   Open a terminal application on the PC and connect to the EV kit's console UART at 115200, 8-N-1.
-
-### MAX32650FTHR Evaluation Kit
--   Connect a USB cable between the PC and the J1 connector.
--   Connect a debugger to J5.
--   Open a terminal application on the PC and connect to the EV kit's console UART at 115200, 8-N-1.
-
-## Expected Output
-
-The Console UART of the device will output these messages:
-
-```
-********************** Static Library Example **********************
-
-This example uses static library functions to read the state of SW1 and
-SW2 buttons then toggles RED and GREEN LEDs according to the values read.
-
-
-...
-```
-
