--- Examples/MAX32650/TestLib/.vscode/launch.json
@@ -1,133 +1,0 @@
-{
-    "configurations": [
-        {
-            "name": "Debug Arm (Cortex-debug)",
-            "cwd":"${workspaceRoot}",
-            "executable": "${workspaceFolder}/build/${config:program_file}",
-            "loadFiles": ["${workspaceFolder}/build/${config:program_file}"],
-            "symbolFiles": [{
-                "file": "${workspaceFolder}/build/${config:symbol_file}"
-            }],
-            "request": "launch",
-            "type": "cortex-debug",
-            "servertype": "openocd",
-            "linux": {
-                "gdbPath": "${config:ARM_GCC_path}/bin/arm-none-eabi-gdb",
-                "serverpath": "${config:OCD_path}/openocd",
-            },
-            "windows": {
-                "gdbPath": "${config:ARM_GCC_path}/bin/arm-none-eabi-gdb.exe",
-                "serverpath": "${config:OCD_path}/openocd.exe",
-            },
-            "osx": {
-                "gdbPath": "${config:ARM_GCC_path}/bin/arm-none-eabi-gdb",
-                "serverpath": "${config:OCD_path}/openocd",
-            },
-            "searchDir": ["${config:OCD_path}/scripts"],
-            "configFiles": ["interface/${config:M4_OCD_interface_file}", "target/${config:M4_OCD_target_file}"],
-            "interface": "swd",
-            "runToEntryPoint": "main",
-            "svdFile": "${config:MAXIM_PATH}/Libraries/CMSIS/Device/Maxim/${config:target}/Include/${config:target}.svd"
-        },
-        {
-            "name": "GDB (Arm M4)",
-            "type": "cppdbg",
-            "request": "launch",
-            "program": "${workspaceFolder}/build/${config:program_file}",
-            "args": [],
-            "stopAtEntry": true,
-            "cwd": "${workspaceFolder}",
-            "environment": [],
-            "externalConsole": false,
-            "MIMode": "gdb",
-            "linux": {
-                "miDebuggerPath": "${config:ARM_GCC_path}/bin/arm-none-eabi-gdb",
-                "debugServerPath": "${config:OCD_path}/openocd",
-            },
-            "windows": {
-                "miDebuggerPath": "${config:ARM_GCC_path}/bin/arm-none-eabi-gdb.exe",
-                "debugServerPath": "${config:OCD_path}/openocd.exe",
-            },
-            "osx": {
-                "miDebuggerPath": "${config:ARM_GCC_path}/bin/arm-none-eabi-gdb",
-                "debugServerPath": "${config:OCD_path}/bin/openocd",
-            },
-            "logging": {
-                "exceptions": true,
-                "trace": false,
-                "traceResponse": false,
-                "engineLogging": false
-            },
-            "miDebuggerServerAddress": "localhost:3333",
-            "debugServerArgs": "-s ${config:OCD_path}/scripts -f interface/${config:M4_OCD_interface_file} -f target/${config:M4_OCD_target_file} -c \"init; reset halt\"",
-            "serverStarted": "Info : Listening on port 3333 for gdb connections",
-            "filterStderr": true,
-            "targetArchitecture": "arm",
-            "customLaunchSetupCommands": [
-                {"text":"-list-features"}
-            ],
-            "setupCommands": [
-                { "text":"set logging overwrite on"},
-                { "text":"set logging file debug-arm.log"},
-                { "text":"set logging on"},
-                { "text":"cd ${workspaceFolder}" },
-                { "text":"exec-file build/${config:program_file}" },
-                { "text":"symbol-file build/${config:symbol_file}" },
-                { "text":"target remote localhost:3333" },
-                { "text":"monitor reset halt" },
-                { "text":"set $pc=Reset_Handler"},
-                { "text":"b main" }
-            ]
-        },
-        {
-            "name": "GDB (RISC-V)",
-            "type": "cppdbg",
-            "request": "launch",
-            "program": "${workspaceFolder}/buildrv/${config:program_file}",
-            "args": [],
-            "stopAtEntry": false,
-            "cwd": "${workspaceFolder}",
-            "environment": [],
-            "externalConsole": false,
-            "MIMode": "gdb",
-            "linux": {
-                "miDebuggerPath": "${config:xPack_GCC_path}/bin/riscv-none-embed-gdb",
-                "debugServerPath": "${config:OCD_path}/openocd",
-            },
-            "windows": {
-                "miDebuggerPath": "${config:xPack_GCC_path}/bin/riscv-none-embed-gdb.exe",
-                "debugServerPath": "${config:OCD_path}/openocd.exe",
-            },
-            "osx": {
-                "miDebuggerPath": "${config:xPack_GCC_path}/bin/riscv-none-embed-gdb",
-                "debugServerPath": "${config:OCD_path}/bin/openocd",
-            },
-            "logging": {
-                "exceptions": true,
-                "trace": false,
-                "traceResponse": false,
-                "engineLogging": false
-            },
-            "miDebuggerServerAddress": "localhost:3334",
-            "debugServerArgs": "-c \"gdb_port 3334\" -s ${config:OCD_path}/scripts -f interface/${config:RV_OCD_interface_file} -f target/${config:RV_OCD_target_file}",
-            "serverStarted": "Info : Listening on port 3334 for gdb connections",
-            "filterStderr": true,
-            "customLaunchSetupCommands": [
-                {"text":"-list-features"}
-            ],
-            "targetArchitecture": "arm",
-            "setupCommands": [
-                { "text":"set logging overwrite on"},
-                { "text":"set logging file debug-riscv.log"},
-                { "text":"set logging on"},
-                { "text":"cd ${workspaceFolder}" },
-                { "text": "set architecture riscv:rv32", "ignoreFailures": false },          
-                { "text":"exec-file build/${config:program_file}", "ignoreFailures": false },
-                { "text":"symbol-file buildrv/${config:symbol_file}", "ignoreFailures": false },
-                { "text":"target remote localhost:3334" },
-                { "text":"b main" },
-                { "text": "set $pc=Reset_Handler","ignoreFailures": false }
-            ]
-        }
-    ]
-}
